+++ src/index.ts
@@ -331,6 +331,7 @@ export {
 	validateForCodeSigning,
 	validateForTlsClient,
 	validateForTlsServer,
+	VERIFY_ERROR_CODES,
 	verifyCertificateChain,
 	verifyCertificateSigningRequest,
 } from './verify/index.ts';